Moved instances of menus to controller
