+++ GUIController.py
@@ -1,6 +1,7 @@
 import sys
 from ClassCollection import ClassCollection
 from PyQt5.QtWidgets import QLabel
+import GUIController, GUIClassMenu, GUIRelationshipMenu, GUIFieldMenu, GUIParameterMenu, GUIMethodMenu, GUIBrowseFiles
 
 # A default collection
 collection = ClassCollection()
@@ -67,6 +68,58 @@ class GUIController:
 
     def renameField(self, className, oldName, newName):
         pass
+
+    # Creates instance of popup windows
+    def addClassMenu(self, checked):
+        self.cMenu = GUIClassMenu.ClassMenu().addClass()
+ 
+    def addRelationshipMenu(self, checked):
+        self.rMenu = GUIRelationshipMenu.RelationshipMenu().addRelationship()
+
+    def addFieldMenu(self, checked):
+        self.fMenu = GUIFieldMenu.FieldMenu().addField()
+    
+    def addMethodMenu(self, checked):
+        self.mMenu = GUIMethodMenu.MethodMenu().addMethod()
+
+    def addParamMenu(self, checked):
+        self.pMenu = GUIParameterMenu.ParameterMenu().addParameter()
+
+    def openMenu(self, checked):
+        self.oMenu = GUIBrowseFiles.BrowseFiles().openFile()
+
+    def saveMenu(self, checked):
+        self.sMenu = GUIBrowseFiles.BrowseFiles().saveFile()
+
+    def delClassMenu(self, checked):
+        self.delCMenu = GUIClassMenu.ClassMenu().deleteClass()
+
+    def renClassMenu(self, checked):
+        self.renCMenu = GUIClassMenu.ClassMenu().renameClass()
+
+    def delFieldMenu(self, checked):
+        self.delFMenu = GUIFieldMenu.FieldMenu().deleteField()
+
+    def renFieldMenu(self, checked):
+        self.renFMenu = GUIFieldMenu.FieldMenu().renameField()
+
+    def delMethodMenu(self, checked):
+        self.delMethodMenu = GUIMethodMenu.MethodMenu().deleteMethod()
+
+    def renMethodMenu(self, checked):
+        self.renMethodMenu = GUIMethodMenu.MethodMenu().renameMethod()
+
+    def delParamMenu(self, checked):
+        self.delPMenu = GUIParameterMenu.ParameterMenu().deleteParameter()
+
+    def renParamMenu(self, checked):
+        self.renPMenu = GUIParameterMenu.ParameterMenu().renameParameter()
+
+    def delRelationshipMenu(self, checked):
+        self.delRMenu = GUIRelationshipMenu.RelationshipMenu().deleteRelationship()
+
+    def renRelationshipMenu(self, checked):
+        self.renRMenu = GUIRelationshipMenu.RelationshipMenu().renameRelationship()
 
 
 # Create a method to pull info about classes from a class collection instance. Maybe that should be made here?
